requirements.txt
Django==1.11.3
django-rest-framework==0.1.0
djangorestframework==3.6.3
psycopg2==2.7.3
py==1.4.34
Pygments==2.2.0
pytest==3.1.3
pytest-django==3.1.2
pytest-watch==4.1.0
pytz==2017.2
